refactor: remove client logic of layout
