+++ app/(main)/layout.tsx
@@ -1,43 +1,11 @@
-'use client';
-
-import { ReactNode, useEffect } from 'react';
-import { Button, Pagination, ResultList, Snackbar } from '@/app/ui/components';
-import { useGetSearchAnimeQuery } from '@/app/lib/utils/animeApi';
-import { useSearchParams } from 'next/navigation';
-import { useQueryUpdater } from '@/app/lib/hooks/useQueryUpdater';
+import { ReactNode, Suspense } from 'react';
+import HomeWrapper from '../ui/home/home-wrapper';
+import { Loading } from '../ui/components';
 
 export default function MainSectionLayout({ children }: { children: ReactNode }) {
-  const searchParams = useSearchParams();
-  const searchTerm = searchParams.get('query') ?? '';
-  const page = searchParams.get('page') ?? '1';
-  const { data, error, isFetching, refetch } = useGetSearchAnimeQuery({ searchTerm, page });
-  const { updateParam } = useQueryUpdater();
-
-  useEffect(() => {
-    if (!searchParams.get('page')) {
-      updateParam('page', '1');
-    }
-  }, [searchParams, updateParam]);
-
-  const handleNumberPage = (page: number) => {
-    updateParam('page', page.toString());
-  };
-
-  const handleRefresh = () => {
-    refetch();
-  };
-
   return (
-    <>
-      <div className="flex flex-col flex-1 gap-4 justify-between min-w-[35%] max-w-[35%] min-h-[85vh]">
-        <Button disabled={isFetching} onClick={handleRefresh}>
-          {isFetching ? 'Refreshing ...' : 'Refresh'}
-        </Button>
-        <ResultList data={data?.data} error={error} loading={isFetching} />
-        {data?.pagination && <Pagination pagination={data.pagination} handleNumberPage={handleNumberPage} />}
-      </div>
-      {children}
-      <Snackbar />
-    </>
+    <Suspense fallback={<Loading />}>
+      <HomeWrapper>{children}</HomeWrapper>
+    </Suspense>
   );
 }
