+++ app/about/page.tsx
@@ -1,6 +1,9 @@
-'use client';
-
+import { Metadata } from 'next';
 import Link from 'next/link';
+
+export const metadata: Metadata = {
+  title: 'About',
+};
 
 export default function About() {
   return (
