+++ app/loading.tsx
@@ -1,0 +1,5 @@
+import { Loading } from '@/app/ui/components';
+
+export default function StartLoading() {
+  return <Loading />;
+}
