+++ app/ui/components/header.tsx
@@ -2,6 +2,8 @@ import { Suspense } from 'react';
 import Navigation from './navigation';
 import Search from './search';
 import { Loading } from './loading/loading';
+import Link from 'next/link';
+import { ThemeToggle } from './theme-toggle/theme-toggle';
 
 const CONTENT = {
   title: 'Anime searcher',
@@ -11,10 +13,13 @@ export const Header = () => {
   return (
     <header
       className={`flex justify-between items-center gap-5 px-5 py-2.5 border-b-2 border-amber-100 dark:border-gray-600'`}>
-      <h3 className="text-3xl font-bold text-orange-400 [text-shadow:2px_2px_4px_rgba(0,0,0,0.7)]">{CONTENT.title}</h3>
+      <h3 className="text-4xl font-bold cursor-pointer text-orange-600 text-shadow-amber-950">
+        <Link href={'/'}>{CONTENT.title}</Link>
+      </h3>
       <Suspense fallback={<Loading />}>
         <Search />
         <Navigation />
+        <ThemeToggle />
       </Suspense>
     </header>
   );
