+++ next.config.mjs
@@ -2,6 +2,9 @@
 const nextConfig = {
   distDir: './dist',
   devIndicators: false,
+  images: {
+    remotePatterns: [new URL('https://cdn.myanimelist.net/**')],
+  },
 };
 
 export default nextConfig;
