+++ package.json
@@ -30,7 +30,6 @@
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-redux": "^9.2.0",
-    "react-router": "^7.7.1",
     "tailwindcss": "^4.1.11"
   },
   "devDependencies": {
